+++ R/stability.R
@@ -1,24 +1,24 @@
-#' Cophenetic distance between leaves of unweighted tree
+#' Graph Geodesic between leaves of unweighted tree
 #'
 #' @param x Object of class `phylo`.
 #' @param nTip Integer specifying number of leaves.
 #' @param asMatrix Logical specifying whether to coerce output to matrix format.
-#' @return `Cophenetic()` returns an unnamed integer matrix describing the
+#' @return `GraphGeodesic()` returns an unnamed integer matrix describing the
 #' number of edges between each pair of edges.
 #' @author Martin R. Smith, modifying algorithm by Emmanuel Paradis
 #' in `ape::dist.nodes()`.
 #' @importFrom ape dist.nodes
 #' @keywords internal
 #' @examples
-#' Cophenetic(TreeTools::BalancedTree(5))
+#' GraphGeodesic(TreeTools::BalancedTree(5))
 #' @useDynLib Rogue, .registration = TRUE
 #' @export
-Cophenetic <- function (x, nTip = length(x$tip.label), log = FALSE,
-                        asMatrix = TRUE) {
+GraphGeodesic <- function (x, nTip = length(x$tip.label), log = FALSE,
+                           asMatrix = TRUE) {
   x <- Preorder(x)
   edge <- x$edge - 1L
   nNode <- x$Nnode
-  ret <- .Call(if (isTRUE(log)) "COPHENETIC_LOG" else "COPHENETIC",
+  ret <- .Call(if (isTRUE(log)) "LOG_GRAPH_GEODESIC" else "GRAPH_GEODESIC",
                       n_tip = as.integer(nTip),
                       n_node = as.integer(nNode),
                       parent = as.integer(edge[, 1]),
@@ -37,18 +37,40 @@ Cophenetic <- function (x, nTip = length(x$tip.label), log = FALSE,
   }
 }
 
+#' @rdname GraphGeodesic
+#' @export
+Cophenetic <- function (x, nTip = length(x$tip.label), log = FALSE,
+                        asMatrix = TRUE) {
+  .Deprecated('GraphGeodesic')
+  GraphGeodesic(x, nTip, log, asMatrix)
+}
+
 #' Tip instability
 #'
 #' `TipInstability()` calculates the instability of each leaf in a tree.
 #' Unstable leaves are likely to display roguish behaviour.
 #'
 #' \insertCite{SmithCons;textual}{Rogue} defines the *instability* of a pair
-#' of leaves as the median absolute divergence in the cophenetic distance
+#' of leaves as the median absolute divergence in the graph geodesic
 #' (the number of edges in the shortest path between the leaves) across all
-#' trees, normalized against the mean cophenetic distance.
+#' trees, normalized against the mean graph geodesic.
 #' The instability of a single leaf is the mean instability of all pairs that
 #' include that leaf; higher values characterise leaves whose position is more
 #' variable between trees.
+#'
+#' Other concepts of leaf instability include
+#'
+#' - The 'taxonomic instability index', as implemented in Mesquite:
+#' described by \insertCite{Thomson2010;textual}{Rogue} as
+#' $\sum\limits_{(x, y), j \neq i}{\frac{|D~ijx~ - D~ijy~|}{(D~ijx~ - D~ijy~)^2}}$,
+#' where $D~ijx~$ is the patristic distance (i.e. length of edges) between
+#' leaves $i$ and $j$ in tree $x$.
+#'
+#' - the average stability of triplets (i.e. quartets including the root) that
+#' include the leaf \insertCite{Thorley1999}{Rogue}, implemented in Phyutility
+#' \insertCite{Smith2008}{Rogue}.
+#'
+#' - @Thomson2010
 #'
 #' @inheritParams RogueTaxa
 #' @param log Logical specifying whether to log-transform distances when
@@ -94,7 +116,7 @@ TipInstability <- function (trees, log = TRUE, average = 'mean',
   }
   nTree <- length(trees)
 
-  dists <- vapply(trees, Cophenetic, double(nTip * nTip),
+  dists <- vapply(trees, GraphGeodesic, double(nTip * nTip),
                   nTip = nTip, log = log, asMatrix = FALSE)
 
   whichDev <- pmatch(tolower(deviation), c('sd', 'mad'))